// projects/notado-lib/src/lib/util/component-type.enum.ts
/**
 * !! When adding new type add it also in BE side ComponentTypeEnum.java
 */
export enum ComponentTypeEnum {
    TEXT = 'TEXT',
    CHECKBOX = 'CHECKBOX',
    LABEL = 'LABEL',
    LINE = 'LINE',
    TEXTAREA = 'TEXTAREA',
    COLOR = 'COLOR',
    DATEPICKER = 'DATEPICKER',
    TIMEPICKER = 'TIMEPICKER',
    ATTENDEES = 'ATTENDEES',


}